Moved check_outputs to check package
