+++ cmd/process-agent/collector.go
@@ -7,7 +7,6 @@ package main
 
 import (
 	"fmt"
-	"github.com/DataDog/datadog-agent/pkg/process/exports"
 	"math/rand"
 	"net/http"
 	"strconv"
@@ -157,7 +156,7 @@ func (l *Collector) runCheck(c checks.Check, results *api.WeightedQueue) {
 		log.Errorf("Unable to run check '%s': %s", c.Name(), err)
 		return
 	}
-	exports.StoreCheckOutput(c.Name(), messages)
+	checks.StoreCheckOutput(c.Name(), messages)
 	l.messagesToResults(start, c.Name(), messages, results)
 
 	if !c.RealTime() {
@@ -178,13 +177,13 @@ func (l *Collector) runCheckWithRealTime(c checks.CheckWithRealTime, results, rt
 	}
 	l.messagesToResults(start, c.Name(), run.Standard, results)
 	if options.RunStandard {
-		exports.StoreCheckOutput(c.Name(), run.Standard)
+		checks.StoreCheckOutput(c.Name(), run.Standard)
 		logCheckDuration(c.Name(), start, runCounter)
 	}
 
 	l.messagesToResults(start, c.RealTimeName(), run.RealTime, rtResults)
 	if options.RunRealTime {
-		exports.StoreCheckOutput(c.RealTimeName(), run.RealTime)
+		checks.StoreCheckOutput(c.RealTimeName(), run.RealTime)
 	}
 }
